refactor(automation): own combined admission and terminal policy
